backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Bcrypt
psycopg2-binary
pytest
